Added confirm modal

+++ components/data-table/DataTable.tsx
@@ -19,6 +19,8 @@ import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
 import { Trash } from 'lucide-react'
 
+import { useConfirm } from '@/hook/use-confirm'
+
 interface DataTableProps<TData, TValue> {
   columns: ColumnDef<TData, TValue>[]
   data: TData[]
@@ -34,6 +36,10 @@ export function DataTable<TData, TValue>({
   onDelete,
   disabled
 }: DataTableProps<TData, TValue>) {
+  const [ConfirmationDialog, confirm] = useConfirm(
+    'Are you sure you want to delete the selected accounts?',
+    'This action cannot be undone.'
+  )
   const [sorting, setSorting] = React.useState<SortingState>([])
   const [columnFilters, setColumnFilters] = React.useState<ColumnFiltersState>([])
   const [rowSelection, setRowSelection] = React.useState({})
@@ -57,6 +63,7 @@ export function DataTable<TData, TValue>({
 
   return (
     <div>
+      <ConfirmationDialog />
       <div className='flex items-center py-4'>
         <Input
           placeholder={`Filter ${filterKey}...`}
@@ -70,9 +77,13 @@ export function DataTable<TData, TValue>({
             variant='outline'
             size='sm'
             disabled={disabled}
-            onClick={() => {
-              onDelete(table.getFilteredSelectedRowModel().rows)
-              table.resetRowSelection()
+            onClick={async () => {
+              const ok = await confirm()
+
+              if (ok) {
+                onDelete(table.getFilteredSelectedRowModel().rows)
+                table.resetRowSelection()
+              }
             }}
           >
             <Trash className='size-4 mr-2' />
